Added Main Python File and the key file
This script runs off of a single python file currently. I am pretty new to coding so its basically spaghetti. The key.key file is necessary for the fernet encryption that I use. Its really basic and easily reversible as long as you have the key and a basic understanding of how it works. If you want a different key, then make one yourself or maybe I will add some code on startup that checks for a key.key file and creates a new one if it is not found. :)